Side-scrolling, random loot, item tweaks and more
- added jump boost, invincibility, carry weight, health boost, and
strength potions
- added damage/defense multipliers to player, as well as a maximum
health value
- fixed fallSpeed not resetting when hitting bottom of platform
- right clicking on level objects is disabled when you have a ranged
weapon in your hand
- changed the "shoot" key from left to right click
- added lots of new crafting recipes and ingredients
- storage menus now center the item images in the slots
- added an options page (stores settings in the .labrunner folder)
- rewrote crafting code; now recipes are added to a database and the
crafting GUI compares your items to each recipe
- Pressing CONTROL-Q will drop your entire inventory
- added basic scrolling through hotbar (will finish it later)
- added more methods to Mouse and Screen class
- changed the look of the titlescreen; added more buttons and changed
the background
- some buttons will only perform an action when the mouse is released
now
- only one instance of UnicodeFont is loaded (instead of one per menu)
- improved accuracy of projectiles (tweaked the getAngleOfElevation()
method)
- added sidescrolling support (bigger levels!)
- added a null check to the rotateImage() method
- merged all items into one class; items are spawned by calling "new
ITEMTYPE(name, id, x, y)"
- removed crafting objects in favor of pressing C to open the crafting
menu
- chests can now randomly generate loot
- merged levels into one class
- added the ability to save your game
- added black holes
- added ammunition crates that add ammo to the currently selected weapon
(if you have not exceeded your ammo limit)

+++ powerup/ammunition/Ammunition.java
@@ -3,14 +3,36 @@ package powerup.ammunition;
 import database.ObjectList;
 import enemy.Scientist;
 import item.Item;
+import item.projectiles.Projectile;
 import item.weapons.Weapon;
 import org.newdawn.slick.Graphics;
+import org.newdawn.slick.Image;
+import org.newdawn.slick.SlickException;
 import player.Inventory;
 import powerup.PowerUp;
 
 public class Ammunition extends PowerUp {
     
+    String targetWeapon; //the weapon that the ammo pack will refill
+    
+    public Ammunition(double x, double y) {
+        
+        this.X = x;
+        this.Y = y;
+        
+        try {
+            this.defaultTexture = new Image("./resources/placeholder.png", false, Image.FILTER_NEAREST);
+        } catch (SlickException ex) {
+            ex.printStackTrace();
+        }
+        
+        ObjectList.powerups.add(this);
+    }
+    
     public void update() {
+        
+        W = defaultTexture.getWidth();
+        H = defaultTexture.getHeight();
         
         hitbox.setBounds((int) X, (int) Y, W, H);
         topHitbox.setBounds((int) X, (int) Y, W, H / 3);
@@ -26,8 +48,25 @@ public class Ammunition extends PowerUp {
     }
     
     public void affect(Object target) {
-        ((Weapon)target).ammoAmount += strength;
-        this.delete();
+        
+        System.out.println(target);
+
+        try {
+            if (Inventory.ammoAmount < 100 && target != null) {
+
+                String a = ((Weapon)target).ammunition;
+                Object aa = new Projectile(a, 9999, 9999, 1, this);
+
+                ((Weapon)target).ammoAmount += 40 / ((Projectile)aa).value; //grant up to 40 ammo points
+                Inventory.ammoAmount += ((Projectile)aa).value * (40 / ((Projectile)aa).value);
+                ((Item)aa).delete();
+                this.delete();
+            } else {
+                System.err.println("Conditions not met; no ammo will be given!");
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
     }
     
     public void draw(Graphics g) {
